+++ example/lib/home.dart
@@ -8,20 +8,39 @@ class HomeScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return SafeArea(
       child: Scaffold(
-        body: Center(
-          child: BubblePopUp(
-            popUpColor: Colors.green,
-            popUp: Container(
-              width: 100,
-              height: 100,
-              color: Colors.green,
-            ),
-            child: Container(
-              height: 200,
-              width: 200,
-              color: Colors.red,
-            ),
-          ),
+        body: Padding(
+          padding: const EdgeInsets.all(20.0),
+          child: PopupScope(
+              builder: (_) => Stack(
+                    children: [
+                      Positioned(
+                        top: 200,
+                        left: 200,
+                        child: BubblePopUp(
+                          config: BubblePopUpConfig(
+                            baseAnchor: Alignment.topCenter,
+                            popUpAnchor: Alignment.bottomLeft,
+                            arrowDirection: ArrowDirection.down,
+                            childBorderRadius: BorderRadius.circular(10),
+                          ),
+                          popUpColor: Colors.green,
+                          popUp: Container(
+                            width: 200,
+                            height: 50,
+                            decoration: BoxDecoration(
+                              color: Colors.green,
+                              borderRadius: BorderRadius.circular(10),
+                            ),
+                          ),
+                          child: Container(
+                            width: 50,
+                            height: 50,
+                            color: Colors.red,
+                          ),
+                        ),
+                      ),
+                    ],
+                  )),
         ),
       ),
     );